requirements.txt
fastapi>=0.111,<1
uvicorn[standard]>=0.29,<1
requests>=2.31,<3
python-dotenv>=1.0,<2
